Created base Skeleton and abstracted Ui and Storage
